Adds a dark mode that auto defaults to system preference. Also changes contact button to an email icon.
